refactor(UserInputView)
- 참여자를 입력받는 InputView와 최대 사다리 높이를 입력받는 InputView 분리
