+++ src/main/java/nextstep/ladder/ui/UiController.java
@@ -8,12 +8,17 @@ import java.util.Scanner;
 
 public class UiController {
     public static void main(String[] args) {
-        Scanner userInputScanner = new Scanner(System.in);
-        InputView userInput = InputView.createByUserInput(userInputScanner);
+        Scanner playerInputScanner = new Scanner(System.in);
+        PlayerInputView playerInputView = PlayerInputView.createByUserInput(playerInputScanner);
+
+        Scanner maxLadderHeightInputScanner = new Scanner(System.in);
+        LadderHeightInputView ladderHeightInputView = LadderHeightInputView
+                .createByUserInput(maxLadderHeightInputScanner);
 
         PointAddStrategy pointAddStrategy = new RandomPointAddStrategy();
-        Ladder ladder = Ladder.create(userInput.getMaxLadderHeight(), userInput.getPlayerSize(), pointAddStrategy);
-        OutputView outputView = new OutputView(userInput.getPlayers(), ladder);
+        Ladder ladder = Ladder.create(ladderHeightInputView.getMaxLadderHeight(),
+                playerInputView.getPlayerSize(), pointAddStrategy);
+        OutputView outputView = new OutputView(playerInputView.getPlayers(), ladder);
 
         System.out.println(outputView.parsePlayerNames());
         System.out.println(outputView.parseLadder());
